ci: improve pipelines (#290)
- separated purging template for main and feature branches
The issue here was that we still tried to purge the version folder even
though we only deployed into a feature branch folder. As the version
variables haven't been set (in case we have a feature branch), the
purging failed. I separated the template and added the `$branch`
variable as folder path for the purging step.

- corrected syntax in condition for deployment type "none"
As the condition in the `deploy-to-cdn.yml` on line 30-41 works and also
the one from line 82-88 works fine, I assume that we just missed the
"()" around the `$deployment` variable. The install and build step was
running everytime even though we just had delpoyment type "none". lets
see if it works

- added more logs to easier see what happens (or not) in each step
I just added some more logs to see that nothing happed in the "else"
case because of deployment type "none".