docs: 📚 change defaults for sd-teaser (#358)
### Description:
I changed the default to 'white border-neutral-300' everywhere but the
'Inset and Orientation Story', since the inset is always set to true in
this variant so it wouldn't show the behaviour. Instead I left it on
default 'white', but the user can switch between variants in the
controls to see how any of them looks like. If you have other
suggestions let me know.

Related to #349