fix: remove automatic valid styling from form fields
BREAKING CHANGE: Before this change `sd-input`, `sd-select` and `sd-textarea` immediately showed valid styles (success color + checkmark) when an form field was valid. In most cases this isn't relevant, e. g. in search fields, app interfaces etc.
With this change you now have to explicitly set the attribute `style-on-valid` on the mentioned components to show "valid" styles, as soon as the component is valid.