fix: sd-tooltip self closing with click focus trigger (close #1211) (#1220)
<!-- ## Title: Please consider adding the [skip chromatic] flag to the
PR title in case you dont need chromatic testing your changes. -->
## Description:
If the tooltip trigger is set to `click focus`, the focus event will be
triggered first and the attached logic will run. Afterwards, the logic
of the `click`event will also run, meaning that the tooltip will open
and close immediately. In this PR we add a flag that will control which
of the event handle should be executed.

## Definition of Reviewable:
<!-- *PR notes: Irrelevant elements should be removed.* -->
- [ ] Documentation is created/updated
- [ ] Migration Guide is created/updated
- [x] E2E tests (features, a11y, bug fixes) are created/updated
- [ ] Stories (features, a11y) are created/updated
- [x] relevant tickets are linked