fix: improve table template a11y (#1664)
<!-- ## Title: Please consider adding the [skip chromatic] flag to the
PR title in case you dont need chromatic testing your changes. -->
## Description:
Closes #1534

Goals:
- [x] Tables without interactive elements are not reachable via
keyboard. Wrap tables with `role="region"`and `tabindex="0"`
- [x] Usage of `role="status"` to announce to screen readers when sort
status has changed.
- [x] Implement `sr-only` `thead` to tables with fixed left, right and
bottom headers.

---

`role="region"` identifies a significant document area, a part from
`role="group"` which is not so significant and is even used on areas
that are not intended to be included in a page summary.

>The region role is used to identify document areas the author deems
significant. It is a generic landmark available to aid in navigation
when none of the other landmark roles are appropriate.

>The group role identifies a set of user interface objects that is not
intended to be included in a page summary or table of contents by
assistive technologies.

## Definition of Reviewable:
<!-- *PR notes: Irrelevant elements should be removed.* -->
- [x] Documentation is created/updated
- [x] Stories (features, a11y) are created/updated
- [x] relevant tickets are linked

---------

Co-authored-by: Karl Baumhauer <48211526+karlbaumhauer@users.noreply.github.com>