feat: disabled form elements (#1899)
<!-- ## Title: Please consider adding the [skip chromatic] flag to the
PR title in case you dont need chromatic testing your changes. -->
## Description:
This PR introduces the `visually-disabled` attribute to form elements so
that they can be shown has disabled but still accessible to screen
readers.
The `help-text` attribute and slot have also been added in the
`sd-radio-group` and `sd-checkbox-group`.

## Definition of Reviewable:
<!-- *PR notes: Irrelevant elements should be removed.* -->
- [ ] Documentation is created/updated
- [ ] Migration Guide is created/updated
- [ ] E2E tests (features, a11y, bug fixes) are created/updated
<!-- *If this PR includes a bug fix, an E2E test is necessary to verify
the change. If the fix is purely visual, ensuring it is captured within
our chromatic screenshot tests is sufficient.* -->
- [ ] Stories (features, a11y) are created/updated
- [ ] relevant tickets are linked

---------

Co-authored-by: Aurora Vasconcelos <aurora.vasconcelos@virtual-identity.com>