fix: sd-copyright shadow not rendering (#2025)
<!-- ## Title: Please consider adding the [skip chromatic] flag to the
PR title in case you dont need chromatic testing your changes. -->
## Description:
This PR fixes the shadow of `sd-copyright` which was not rendering due
to missing variables.
The issue is only visible outside of storybook (eg: codepen) and was
reported in the support channel.

## Definition of Reviewable:
<!-- *PR notes: Irrelevant elements should be removed.* -->
- [ ] Documentation is created/updated
- [ ] Migration Guide is created/updated
- [ ] E2E tests (features, a11y, bug fixes) are created/updated
<!-- *If this PR includes a bug fix, an E2E test is necessary to verify
the change. If the fix is purely visual, ensuring it is captured within
our chromatic screenshot tests is sufficient.* -->
- [ ] Stories (features, a11y) are created/updated
- [ ] relevant tickets are linked

---------

Co-authored-by: mariohamann <mario.hamann@virtual-identity.com>