fix: styles package npm release (#2258)
<!-- ## Title: Please consider adding the [skip chromatic] flag to the
PR title in case you dont need chromatic testing your changes. -->
## Description:
Last December when the styles package was first released, the initial
version was set to 1.4.0 by mistake. It was unpublished from NPM but
since it's there, now that the real 1.4.0 needed to be released, it's
not possible. This PR bumps the package version to 1.4.1 to ensure the
release on NPM and also adds this information to the changelog.

## Definition of Reviewable:
<!-- *PR notes: Irrelevant elements should be removed.* -->
- [ ] Documentation is created/updated
- [ ] Migration Guide is created/updated
- [ ] E2E tests (features, a11y, bug fixes) are created/updated
<!-- *If this PR includes a bug fix, an E2E test is necessary to verify
the change. If the fix is purely visual, ensuring it is captured within
our chromatic screenshot tests is sufficient.* -->
- [ ] Stories (features, a11y) are created/updated
- [ ] relevant tickets are linked